ingestion/requirements.txt
openai==0.27.8
beautifulsoup4==4.12.2
requests==2.31.0
tiktoken==0.4.0
feedparser==6.0.10
pinecone-client==2.2.2
yapf==0.40.0
isort==5.12.0
graphqlclient==0.2.4
feedgen==0.9.0
mindsdb_sdk==1.0.4
